Rename `agsgardex.d.ts` -> `asgardex.ts`
just to have "clean" imports
