+++ src/views/pools/PoolsOverview.tsx
@@ -29,7 +29,7 @@ import * as stakeRoutes from '../../routes/stake'
 import * as swapRoutes from '../../routes/swap'
 import { SwapRouteParams } from '../../routes/swap'
 import { PoolsState } from '../../services/midgard/types'
-import { Maybe, Nothing } from '../../types/asgardex.d'
+import { Maybe, Nothing } from '../../types/asgardex'
 import { PoolDetailStatusEnum } from '../../types/generated/midgard'
 import View from '../View'
 import { ActionColumn, TableAction, BlockLeftLabel } from './PoolsOverview.style'
